feat: add generic realm diff callback
